chore: rename AbstractValidator to AbstractPlugin
